+++ DirectXGameEngine/FXAA.h
@@ -5,7 +5,7 @@
 class FXAA : public PostProcessBase
 {
 public:
-	FXAA(float width, float height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
+	FXAA(UINT width, UINT height, ID3D11Device* _device, ID3D11DeviceContext * _deviceContext, ID3D11SamplerState* _sampler);
 	void setBlurAmount(int in);
 	virtual SRV* draw(SRV* ppSRV);
 	~FXAA();
